Seperate out all the search settings to the search settings entity. Optompize the search to only search enabled controllers.
